Rearranged original old crosstalk script as a task. Ready to test single sensor xtalk task for bugs.
