+++ scripts/sensor_crosstalk.py~
@@ -1,0 +1,11 @@
+import os # maybe just import join?
+import argparse
+
+
+def main(sensor_id, calib_dir, output_dir='./'):
+
+    bias_frame = os.path.join(calib_dir, '{0}_superbias.fits'.format(sensor_id))
+    dark_frame = os.path.join(calib_dir, '{0}_superdark.fits'.format(sensor_id))
+    gains = gain_results.get_amp_gains(sensor_id)
+
+
